requirements.txt
asgiref==3.2.7
Django==3.0.6
pytz==2020.1
sqlparse==0.3.1
django-star-ratings==0.9.1
django-allauth==0.42.0
django-dbbackup==3.3.0
django_cron==0.5.1
